refactor: extract remaining routes and delete routes.py
Extract dashboard, playlists, sync, and account routes into dedicated
modules using FastAPI dependency injection. Delete the original
routes.py (4890 lines at start of refactor, 1132 at deletion).

Auth behavior change: songkick/concert-archives partials now return
307 redirect to /login for unauthenticated users (previously returned
200 with empty HTML). Consistent with the DI pattern used across all
extracted modules. 5 tests updated to match.
